mcp/requirements.txt
fastmcp
psycopg2-binary
python-dotenv
requests
httpx
openai
neo4j
